Also move loader creation
